Separate file for V1 test types
